New format for player objects
In the PL table, players are assigned elements inside, so if an object
is needed, then you just need to call from the items from it.

So, if i want to spin the player, i can just do:

PL[1].NoteField:spin()

Included on this is also a failsafe (which is subject to change),
where if the requested amount is out of the scope of the arrangement,
then it will just return the first available player on the table.